Consider 403 Forbidden as 404 Not Found
This is due to GCS changing its error response codes.
Here's a snippet:

Currently, if a request is not authorized to access an object or its
bucket (such as anonymous requests on a non-public object), Google Cloud
Storage returns an error code 403 if the object exists, and error code
404 if it does not exist. Once the change is rolled out, the error code
returned will be 403 for both of these cases. If the object does not
exist and the request has listing permission over the bucket, Google
Cloud Storage will still return error 404. There is no change in
behavior for buckets with Website Configuration enabled.